refactor: separate container output processing
